"fixed setup.py and uninstall.py and added windows setup.bat and uninstall.bat. Now project works in a virtual environment"

+++ uninstall.py
@@ -2,7 +2,7 @@
 Run this from project folder from terminal as Administrator
 """
 
-import os, sys, json
+import os, sys
 from Data.constants import CONFIG_JSON, COMMANDS_BAT, COMMANDS_SH, CHECK_PW, PW_OBJECTS, OLD_PW_OBJECTS
 
 
@@ -18,12 +18,12 @@ print(f"Moreover file '{PW_OBJECTS}' will be kept and renamed as '{OLD_PW_OBJECT
 print("in this way pw data will be preserved (but passwords will be encrypted with your previous PW password!)\n")
 
 
-# Ask if path is correct
-print("Make sure the following path is the same of PW folder you want to uninstall")
-ask = input(f"`{os.getcwd()}` is the correct path?\n [Y,n]: ")
-if ask not in "sSyY":
-  print("Run this script in the correct path as administrator")
-  sys.exit()
+# get the full path of this file
+this_file_full_path = os.path.abspath(__file__)
+# get full path of project folder
+main_folder_full_path = os.path.dirname(this_file_full_path)
+# move to that directory
+os.chdir(main_folder_full_path)
 
 
 # delete files `check_pw.txt` and `config.json`
@@ -33,6 +33,11 @@ if os.path.exists(CHECK_PW):
   os.remove(CHECK_PW)
 
 # rename file `pw_objects.json` for not loosing data
+if os.path.exists(OLD_PW_OBJECTS):
+  print(f"ERROR: file {OLD_PW_OBJECTS} already exists.")
+  print("If you want to uninstall PW move or delete that file")
+  sys.exit()
+
 if os.path.exists(PW_OBJECTS):
   os.rename(PW_OBJECTS, OLD_PW_OBJECTS)
 
@@ -40,7 +45,7 @@ if os.path.exists(PW_OBJECTS):
 if os.name == "nt": 
   # commands.bat full path
   # add one " for being sure that paths that contain spaces will be supported (in regedit it will write exactly "path" with one ")
-  commands_bat_full_path_str = f'"{os.path.join(os.getcwd(), COMMANDS_BAT)}"'
+  commands_bat_full_path_str = f'\\"{os.path.join(os.getcwd(), COMMANDS_BAT)}\\"'
 
   # get regedit AutoRun paths Values in Command Processor
   tmp_file_path = "tmp.txt"
@@ -55,19 +60,23 @@ if os.name == "nt":
   for line in lines:
     if "AutoRun" in line:
       # get the list of all paths in value "AutoRun"
-      paths = line[21:].strip().split(" & ")
+      paths = line[21:].strip().split("&")
 
-      # get the list of all paths in value "AutoRun"
-      paths_in_line = line[21:].strip()
-      if paths_in_line == "": # case: no paths
-        paths = []
-      elif " & " in paths_in_line: # case: more than a path
-        paths = paths_in_line.split(" & ")
-      else: # case: a single path
-        paths = [paths_in_line]
+ 
+  # if a path starts and ends with ", we must add \ in front of "
+  for i, path in enumerate(paths):
+    # strip every path: it could contain spaces because we are splitting with "&"
+    # but each path could be separated with " & "
+    path = path.strip()
+    if path.startswith('"') and path.endswith('"'):
+      # add \ in front of each "
+      path = f'\\"{path[1:-1]}\\"'
+    # refresh path in paths list
+    paths[i] = path
+
 
   # remove commands.bat file full path from the values
-  if paths != []:
+  if paths != [""]:
     paths.remove(commands_bat_full_path_str)
 
   # join all paths separated by &
